整理了所有的数据库字段，初步修改完成

+++ src/main/java/cn/stylefeng/guns/core/common/constant/Const.java
@@ -40,12 +40,12 @@ public interface Const {
     /**
      * 管理员id
      */
-    Integer ADMIN_ID = 1;
+    Long ADMIN_ID = 1L;
 
     /**
      * 超级管理员角色id
      */
-    Integer ADMIN_ROLE_ID = 1;
+    Long ADMIN_ROLE_ID = 1L;
 
     /**
      * 接口文档的菜单名
